Organize components into folders and update imports
