+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import { WeatherForm } from './components/weatherform';
+import { WeatherForm } from './components/Weather/WeatherForm';
 
 
 function App() {
