--- src/components/navbartitle.tsx
@@ -1,1 +1,0 @@
-export {}
