--- src/components/weatheritem.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import getWeather from "../fake_api/weather_webfake";
-
-export const WeatherItem = () => {
-    
-}
